backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg[binary]==3.1.13
psycopg2-binary==2.9.10
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
httpx==0.25.2
feedparser==6.0.10
beautifulsoup4==4.12.2
celery==5.3.4
redis==5.0.1
google-genai==0.3.0
python-dateutil==2.8.2